requirements.txt
matplotlib==3.10.1
pandas==2.2.3
scikit_learn==1.6.1
seaborn==0.13.2
shap==0.47.0
streamlit
openai
joblib
python-dotenv
